pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "Heron-42ad"
version = "0.7.1"
authors = [
    {name = "George Dimitriadis", email = "gdimitri@hotmail.com"},
]
description = "Heron Node baed experiment designer"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["Node-based", "experiment"]
license = {text = "MIT License"}
classifiers = [
        'Development Status :: 3 - Alpha',
        'Intended Audience :: Science/Research',
        'License :: OSI Approved :: MIT License',
        'Operating System :: POSIX :: Linux',
        'Operating System :: Microsoft :: Windows',
        'Programming Language :: Python :: 3.9',
        'Programming Language :: Python :: 3.10',
        'Programming Language :: Python :: 3.11'
]
dependencies = ['pyzmq>=20.0',
                  'numpy',
                  'dearpygui >= 1.2',
                  'paramiko',
                  'opencv-python>=4.5',
                  'pandas',
                  'h5py',
                  'tornado',
                  'pynput',
                  'pyserial',
                  'psutil',
]

[project.scripts]
Heron = "Heron.gui.editor:run"

[project.urls]
Homepage = "https://github.com/Heron-Repositories/Heron/tree/main"
Documentation = "https://heron-42ad.readthedocs.io/en/latest/index.html"
Issues = "https://github.com/Heron-Repositories/Heron/issues"
